WT-10667 Implement WT testing on other weak-ordered memory model processors such as Apple M1 (#8936)
Improved the find_cmake.sh script to use the Mac Homebrew version of CMake and CTest if installed (as is installed on the M1 Mac Evergreen instances). Simplified Python version checking by splitting the checks into a Python script. Refactored the Mac Intel and Mac ARM build variants to use a common template.