Refactor how mirror tables are created
When mirroring is enabled, we need to create the mirror first otherwise
threads may start working on the base as soon as it is created while its
mirror may not be created yet. By creating the mirror first, we have the
guarantee that both tables exist when the base is created.