WT-11542 Chunk cache I/O throttling (#9909)
This throttles in a few situations:

* Performing a `__wt_read` for the chunkcache
* When we know that accessing a specific preloaded chunk is likely to
cause
disk activity
* The existing throttle in the block manager was moved closer to the
`__wt_read`.

There are situations where this does not throttle:

* Inserting values to the chunk cache. This could theoretically be a
disk
write, but the OS manages it for us.
* Reading older values from the chunk cache. The operating system might
have
paged this out.

These are fixable with the `mincore()` syscall, but it doesn't seem
necessary to go to that effort just yet.

The throttling capacity itself is a new config item. There's still the
notion of a "total", but the chunk cache allowance is subtracted from
that
(to keep the meaning of "total" sane). We could have bundled it into the
"total" amount as another hard-coded percentage (like the checkpoint
allowance), but the problem here is that (1) there's a hard-coded read
allowance and it's not clear how chunk cache reads/writes should be
further
subtracted from that, and (2) the chunk cache would typically be on a
different disk to WiredTiger's home directory, so a completely separate
config item allows users to tune the IO allowance per-disk.

One limitation here is that chunk cache reads and writes are subtracted
from
the one "chunk cache" IO capacity. This is the simplest scheme that
works,
and can be improved later without breaking compatibility.