WT-10743 Perform rollback to stable using multiple worker threads
The main thread populates the queue with the files that are needed to be
processed by the rollback to stable and each worker thread picks those
files and process them in parallel.

Once all the data store files are finished, the main thread completes
the rollback to stable by processing the history store table.