WT-9420 Run test_checkpoint on tiered tables in Evergreen (#10184)
To test this, I've versions of the `checkpoint-stress-test` with tiered
storage enabled, both on my workstation and in evergreen. I've also run
patch builds to confirm that the new test shows up and is executed
successfully in the correct variant.

I'm open to suggestions of other variants where we should run this test,
but I don't think we need to over do it. I picked Arm because that is
where tiered storage would most likely be deployed.