WT-11626 Add failure injection to the workload generator (#10169)
This PR introduces the "crash" operation to the workload generator,
which intentionally crashes WiredTiger to force it to do recovery—and
thus to run RTS, as the main goal of this project is to add more RTS
testing.

The trick to getting this working with WiredTiger was to run the
workload in a child process, so that we can easily crash and restart it.
The child process saves its state between crashes using shared memory,
and it also uses it to return information back to the controller parent
process, e.g., whether the crash was intentional.

Feel free to let me know if you have any questions, or if you'd like me
to go over anything via Zoom.