WT-12360 Add model unit testing to PR testing (#10217)
The PR adds model testing to PR testing. The rationale is explained in
the ticket, but TL;DR: If changes to WT result in API-visible changes,
e.g., due to a recent change where setting a timestamp can now return
EINVAL, it would result in failing this test—which may not be caught
until after the PR gets merged.