WT-12225 Fix RNG generator weakness around mongodb $sample stage (#10232)
The error happens where the records are exactly the same because two
cursors are opened close together under one thread. In windows machine
clock resolution is low, therefore the numbers ts.tv_nsec and ts.tv_sec
are the same. If we only changed the rnd.z seed, we get into a problem
like the original code, where rnd.w is the same and rnd.z seed is
changing. Therefore the solution used all of the 64 bits from RDTSC.