WT-10807 Skip in-memory deleted pages as part of the tree walk (#9877)
Traversing an in-memory page that contains all the deleted values that
are visible to the current transaction led to an increase in latency due
to the time spent skipping these deleted values.

By saving the aggregated timestamp information in the ref when the page
has all deleted values, this aggregated information can be validated
against the transaction snapshot to skip traversing the page completely
and improve the latency when there are many deleted pages.

The downside of this approach is that the in-memory size of each ref is
increased by 8 more bytes, but this increase shouldn't cause any
problem.