WT-12324 Wait for the prefetch threads to complete on the dhandle for evict exclusive access (#10189)
We were encountering a segfault where prefetch was trying to process a
ref on the queue that had already been evicted and freed. We implemented
a fix where the prefetch worker threads would make it known whenever
they popped a ref from the queue (and were actively working on it). Any
eviction threads (in this case, the exclusive eviction thread) would
have to wait for any prefetch threads actively operating on that ref to
finish their work before evicting the ref.

---------

Co-authored-by: Luke Pearson <luke.pearson@mongodb.com>