WT-12323 Disable prefetch for corrupted tables when using wt utility (#10275)
Originally a few asserts were hitting in `test_verify.py` but we
realised that this was because the `wt verify` utility was running on
tables that had purposely been corrupted. According to previous
PRs/designs, pre-fetching is not expected to work on corrupted pages due
to certain ref/freeing guarantees not being met. As a result, I don't
think it makes much sense for pre-fetching to work with the
`read_corrupt` mode in verify either.

Pre-fetch cannot be disabled with the usual approach of reconfiguring
the connection/session as the `runWt` function in the Python test
framework will open up its own connection and session.

Main Change: Add a parameter to the `wt` utility so that prefetching can
be disabled for `wt verify` when required (i.e. in scenarios with
corrupted tables).