WT-11877 Add configuration for maximum wait time between fsync calls (#9797)
for example: we configure io_capacity.tatal=1M, where fsync background
takes 10 seconds to work, which is too long.
Also, if the io_capacity.tatal‘s value is large(1T), sync background
should takes a long time to work. If there is a write operation, but the
threshold (cap->written < cap->threshold) is not met for a long time,
shold we force sync background ?

---------

Co-authored-by: Will Korteland <will.korteland@mongodb.com>