WT-12545 Import an existing workload file into the model (#10267)
The PR adds the ability to parse workload operations from a string, and
it also adds a new option to `model_test` to re-run an existing workload
that was saved to a file. This is, for example, useful when debugging a
failed model-based test.

The PR also adds `operator==` and `operator!=` to all workload
operations, so that the test can check whether the operations were
parsed correctly.