WT-12168 Implement WT_ACQUIRE_READ marked access load (#10256)
This adds a new macro to WiredTiger which implements the "read-acquire"
semantic for ARMv8.2-A+rcpc and above. A compile time check is added to
emit the instructions or not depending on the hardware. For details see
"rcpc_test.cmake". Additionally a very simple catch2 unit test was
created to ensure that it works for different sized types.