WT-12609 Improve checkpoint cleanup and page eviction logic
Checkpoint reads only internal pages into the cache for the purpose
of removing any obsolete pages as part of the checkpoint cleanup.
Evicting these internal pages asap can add an overhead to the next
checkpoint to read them back again into the cache. Instead of evicting
them asap, evict the internal pages read by the checkpoint like a
regular page.

As checkpoint never reads any leaf page into the cache, there is no
point in evicting the leaf pages that are marked with READ_WONT_NEED.
This will avoid the unnecessary slow down to the checkpoint by evicting
the pages that are not read by the checkpoint.