WT-12616 Expected crashes in the model should not leave behind cores (#10333)
The PR changes how the child process terminates, so that it does not
leave behind a core file if the crash is intentional.

I did not change the method names; they still contain the word "abort"
even though `abort` is no longer used, as I couldn't come up with a
better name—and this is close enough. Feel free to suggest better naming
in the comments :)