WT-11178 Detect bitmap value change for incremental backups (#9437)
When tracking blocks for incremental backup, the set of changed blocks
is tracked via a bitmap in the code. There is an invariant that single
bits should never be cleared from the block tracking bitmap - they
should only be added. This ticket adds checks that work across a
WiredTiger restart to detect any bits incorrectly getting changed from 1
to 0.

---------

Co-authored-by: Ubuntu <ubuntu@ip-10-0-5-65.ec2.internal>