WT-12249 Opening a cursor may accidentally inherit config from a cached cursor (#10172)
This is marked draft because I need the questions answered below. The
questions might not require any changes to the code if my approach is
acceptable.

NOTE: Please, if no changes are needed, do not click "Merge when ready".
The commit message will be wrong. I will squash the commits to one
myself using "git rebase -i HEAD~2" and make sure the commit message of
the remaining commit is correct. Then I can click "Merge when ready". If
there is only one commit, github uses its commit message for the commit
message of the merged commit.
https://github.blog/changelog/2022-08-23-new-options-for-controlling-the-default-commit-message-when-merging-a-pull-request/

All tests pass but there might be changes needed based on the questions
below. The easiest way to get these questions answered was to post this
review.

1. It is not clear to me whether "Opening a cursor may accidentally
inherit config from a cached cursor" is something that can occur now, or
is something that is to be prevented in the future according to issue
comment "... when we add a new configuration to the cursor, we should
add logic to clear the configuration here. But it is often omitted. We
should extract this logic to a separate function to ensure we don't
forget this."

Question: Are the correct bits cleared now? If not, what are the correct
bits to clear?

2.  I show two ways to solve this.
Commit 1. Clear all bits except the bits to be preserved. This might be
more resilient to future changes.
Commit 2. Clear just the bits that need to be cleared either because
they are being set, or because they should always be cleared.

I decided to do the approach in commit 2. If you prefer, I can switch to
the approach in commit 1.

Question: Do you prefer the approach in commit 1 or commit 2? If commit
1, what bits should be preserved? Did I choose the right ones?