WT-12383 Update tcmalloc (#10464)
Update version of tcmalloc bringing it inline with that used by MongoDB.

Remove tcmalloc from CMake build. The build was linking explicitly
against the `tcmalloc` shared object; that is not using the weak symbol
behavior of malloc, etc. This could be fragile as it required explicit
name substitution, and presented the following problems when upgrading:
- The library no longer exports tc_ prefixed versions of malloc, free,
etc.
- The tcmalloc header file now references the C++ <new> header thereby
making incompatible with C.
Pre-existing Evergreen tasks already employed LD_PRELOAD in addition to
explicit linking; and with this change that is the only mechanism
employed.

Add support script for downloading a pre-built cached binary build based
on the current build variant; or if no such pre-cached binary exists:
downloading and building tcmalloc using source code provided by MongoDB.
MongoDB selectively patches tcmalloc so using the github project source
is not suitable at this time.

tcmalloc must be specifically built for use with a sanitizers: TSan,
MSan and ASan; so tcmalloc is removed from Evergreen tasks employing
these sanitizers.

tcmalloc is incompatible with libeatmydata: so this has been removed
from ALL Evergreen tasks.