WT-12819 Allow larger block sizes with wt_binary_decode.py (#10480)
The wt_binary_decode.py tool is hard coded to assume that a block size
larger than 512K indicates a corrupt block – i.e., that the size as
stored in the page header is likely corrupt/garbage if it too big.
MongoDB supports document sizes up to 16MB, so we can legitimately have
blocks at least that big.

This PR increases the cut-off to 17MB, allowing for some slop around a
single large document.