WT-12872 Make the pinned_ratio check in test_chunkcache03 more lenient (#10515)
Occasionally the float division in
`self.assertEqual(total_chunks/pinned_chunks, 2)` would evaluate to
2.0625 instead of 2 and fail the test. This assertion works on the
assumption that all tables in the test have been populated to be exactly
the same size and use the exact same number of chunks, but there is no
guarantee that this is true. Instead round the divided result to handle
small variation is number of chunks used by each table.

This PR also updates the check to test a few more combinations of pinned
tables for certainty.