WT-12708 Avoid null positions from random eviction traversal (#10437) (v5.0 backport)
To address the problem described in the ticket, the solution is to retry
a 100 times until a ref in memory is found.
If not found, then we walk the root page and try to locate the in-mem
ref.
This ensures that we mostly avoid returning NULL from the random descent
function.

After running wtperf and sys perf, including a high-value workload, I've
observed no regression.

---------

Co-authored-by: Siddhartha Mahajan <siddhartha.mahajan@mongodb.com>
(cherry picked from commit 180e5af61e2628799619d5061538ecd96ea6fc7b)