WT-12562 log_prealloc configuration optimization (#10269)
befor this PR:
log_prealloc's type is bool and initialized to 1, In the first second of
pressure testing, if the file_max value is small, a large number of
WiredTigerPreplog.xxxxx files are created by user threads, which may
have some impact on some user requests.

after this PR:
We can create WiredTigerPreplog.xxxxx files in advance，The number of
files can be configured by log_prealloc.

thanks.