WT-12929 Compact to safe free the addr->addr while replacing ref addr (#10546)
Compact operation can rewrite the ref addr to compact the on-disk file.
As part of the rewrite operation, it can free the existing address and
create a new address for this ref on the disk file. During this time, it
is possible that the same ref being accessed in parallel can lead to
access the freed addr->addr. To avoid this problem, safe free the
addr->addr from the compact operation.