WT-12238 Check ref address when skipping pages in compact (#10538)
While traversing entries on an internal page, compact will decide to
checkpoint the tree if it sees pages in the WT_REF_DELETED state as
checkpoint will delete the underlying blocks, freeing up space for
compact to reclaim. However, compact will repeatedly see pages in the
WT_REF_DELETED state even after they are freed as the index in the
parent page is not updated unless a split occurs.

The ref address is freed when we free the underlying blocks. Adding a
check for the ref address here ensures we only proceed with the compact
checkpoint (and another compact pass) if the underlying blocks have not
already been freed.