WT-12934 Address remaining issues with zero length files and git diffs in the Code Change Report (#10552)
[WT-12868](https://jira.mongodb.org/browse/WT-12868) created a
workaround for newly added 0-length files, however @bzm-jas 's PR for
[WT-12837](https://jira.mongodb.org/browse/WT-12837) also has some
deleted 0-length files. Each of these files can generate either a
deleted file record in the diff, or it may generate a record moving that
file to another location as it sees a match in the (0-length) content.

The solution is to also exclude from the diff output file:

- Any deleted files
- Any moved files where the record includes the path of a 0-length file.

---------

Co-authored-by: Andrew Morton <andrew.morton@mongodb.com>