WT-12889 Identify and skip malformed workloads in counterexample reduction (#10553)
Counterexample reduction sometimes produces invalid workloads that do
not actually reproduce the failure of the original workload, but instead
fail because the reduction process itself generated a malformed
workload. So far I've only seen this happening due to producing a
workload with invalid use of timestamps—so check for timestamps early
and skip testing malformed workloads.