WT-12974 Delete unneeded loop in __rec_hs_wrapup() (#10578)
A loop returns 0 if the loop goes all the way to the end. This is not
needed since the following loop loops over the same range, has the same
check within the loop, and also results in return 0 if the first loop
would have gone all the way to the end. The first loop is not
accomplishing anything.

(cherry picked from commit 1989e720deed71f38c35b541a83ecc7c1885d4c6)