WT-12649 Always send in and check remap boolean. (#10591)
In WT-11555 I didn't add in the remap boolean for this code path because
I thought it couldn't race. I was wrong and it too needs the fix from
that ticket. @keitharnoldsmith because you reviewed the other PR.
@raviprakashgiri29 because you've been involved in this ticket. I'm
going to let it run a while, but I was readily reproducing.

(cherry picked from commit 9b8e54f886832840c62954653969df0c013b2c5c)