WT-12999 Add csv to wt_binary_decode with durable history metadata stats (#10593)
This script (indiscriminantly) walks a `wt` file, parsing as much as it
can and printing helpful output for debugging.
Prior to this change it supported printing some information about page +
cell metadata, but this wasn't easily machine readable.

This change adds a `--csv` option which will write the respective stats
per page to a csv file for later analysis.
We are using this to analyse the ratio of durable history metadata vs
key sizes in index files, but there are many other uses for which this
may be helpful.