WT-12733 Decode JSON strings into the original code points (#10551)
The JSON library uses UTF-8 encoding, which has the unfortunate
consequence of encoding all keys and values from the debug log JSON into
UTF-8. As a result, we need to decode keys and values from UTF-8 back to
their original bytes before unpacking them via `data_value::unpack`.

I used the `iconv` library to do the work, because it is standard on
POSIX systems. I would have preferred to use something like `codecvt` if
I could, but it just got deprecated, and I didn't want to add a
dependency on a non-standard library. And no, I didn't find a way to get
our JSON library to use a different encoding; it looks like UTF-8 is
hard-coded in.