WT-12990 Reset background compaction stats when disabled (#10600)
If background compaction did not successfully compact a file, it will
mark it as inefficient and skip it for some time. Even after turning
background compaction off and on, the file won't be checked until the
expiry time has been reached. The behaviour is now different with the
suggested changes: if background compaction is disabled, its stats are
reset and next time background compaction runs, each file will be
investigated again.

Changes:
- New flag `BACKGROUND_COMPACT_CLEANUP_OFF` used when background
compaction is disabled to cleanup its stats.
- Renamed existing flags