WT-12535 Perform obsolete update check as part of reconciliation (#10259)
Searching for obsolete updates whenever a key is modified can take a lot
of time, and not finding any obsolete updates can increase the latency
of the update operation. To improve this situation, limit the obsolete
update checks of a page to only when the page in-memory foot print is
greater than 50% of the maximum page size allowed. To compensate for the
loss of obsolete checks as part of the key modification, perform the
obsolete checks as part of the reconciliation to remove them.

This change can increase the size of obsolete updates present in the
cache, but they will be removed whenever a reconciliation occurs on
those pages.

---------

Co-authored-by: Jonas Bergler <jonas@bergler.name>