WT-13092 Let checkpoint cleanup to remove obsolete time window from pages
The time window that is written to the pages to support MVCC can become
obsolete and it needs to be cleaned to reclaim the disk space occupied
by the obsolete time window.

To support checkpoint cleanup to work on the files that have obsolete time
window informaion, the time aggregate infomration from the checkpoint is
compared again the pinned timestamp. Also, the max newest start/stop durable
timestamp is compared against pinned timestamp for the pages to consider
them as obsolete. Once the pages are identified they are marked for dirty
to let the reconciliation remove the obsolete time window information.
The on-disk pages with obsolete time window information are read into the
cache to reconcile them.

Along with above functionality changes, the following fixes are done:
1. Do not consider 0 as a valid timestamp in oldest start ts calculation.
2. Limit the checkpoint cleanup to 500 pages per btree.