WT-12096 Don't validate timestamp against the on-page value if the update is globally visible (#10595)
If an update is globally visible it can be evicted and EBUSY should not
be returned. Thus the check for EBUSY should be skipped if an update is
globally visible. The bug is the check for EBUSY was not skipped if an
update is globally visible.

Please review just all commits. There were temporary changes that have
been backed out.