WT-8969 Split four eviction statistics into attempts, and failures. (#10644)
Successes = attempts - failures.

If anyone wants to add statistics or logs related to eviction, add them
to this branch,
[wt-8969-eviction-stats](https://github.com/wiredtiger/wiredtiger/tree/wt-8969-eviction-stats).
I will handle the code review, and commit.

If you want to add statistics or logs, you can use this PR to see what
has already be added without becoming a reviewer and with no need to
make comments. You can also add yourself as a reviewer if that is what
you want.