WT-12909 Don't roll back to 0 if the stable timestamp is not set (#10580)
This PR fixes an inconsistency in RTS semantics when the stable
timestamp is not set, which currently depend on whether it is called
directly or during recovery. This change goes with the semantics of RTS
on recovery, which means not rolling back updates based on their durable
timestamps. I confirmed with the relevant teams that MongoDB does not
call RTS directly without the stable timestamp, so this would have no
effect on MongoDB itself. (I also experimented with other RTS semantics,
only to discover that much of our testing depends on not rolling back to
0 on recovery.)

Another reason for making this change is an issue identified in
[WT-12966](https://jira.mongodb.org/browse/WT-12966), which shows that
when the stable timestamp is not set, RTS does not run even if it has
other work to do, e.g., aborting prepared (but uncommitted)
transactions. Once we fix that ticket for RTS to run in those cases, we
would need to ensure that it does not roll back updates based on their
timestamps to preserve the current semantics.

(cherry picked from commit bdf85b1bf462063f50189ab773a4c2e718f0a744)