WT-12917 Fix a race condtion and a memory corruption in __rts_btree_abort_ondisk_kv (#10712)
This PR fixes the following three bugs in RTS:

1. A double-free if adding a tombstone via `__wt_update_serial` fails,
in which case both `__wt_update_serial` and
`__rts_btree_abort_ondisk_kv` would both free the same tombstone. This
was fixed by making the `upd_arg` argument of `__wt_col_modify` and
`__wt_row_modify` to be a double-star pointer; freeing the update would
set the update to NULL, so that the caller would not also try to free
it.
2. An error in `__rts_btree_walk` would leave an acquired hazard pointer
behind. This was fixed by calling `__wt_page_release` during cleanup.
3. A race condition in `__rts_btree_abort_ondisk_kv`, in which case an
RTS pass over the HS could race with `__wti_rec_hs_clear_on_tombstone`.
Please refer to the ticket and the block comment in
`__rts_btree_abort_ondisk_kv` for details.