WT-12096 Don't validate timestamp against the on-page value if the update is globally visible
If an update is globally visible it can be evicted and EBUSY should not
be returned. Thus the check for EBUSY should be skipped if an update is
globally visible. The bug is the check for EBUSY was not skipped if an
update is globally visible.

Revert the revert to put back 3fabafa95bad2f9aee5e861b4b6ed90903ac29d4.

This reverts commit f872e755e3716f30040796bfb299753498206938.