WT-13186 Remove bounded cursor support when positioning truncate cursors (#10825)
This PR removes the use of bounded cursors when positioning cursors for
a truncate call. Only some cursor types implement bounding so we need to
know which cursor type we're using in `__wt_session_range_truncate`, but
this information isn't available at the api layer. Ongoing work has been
split out to into WT-13232 to review if there's an alternative approach
possible.