WT-13241 Mark obsolete time window pages dirty during checkpoint cleanup
The time window that is written to the pages to support MVCC can become
obsolete and it needs to be cleaned to reclaim the disk space occupied
by the obsolete time window.

As part of the checkpoint cleanup that is going through the in-memory
obsolete time window pages, mark them dirty to let the reconciliation to
remove the obsolete time window from the pages.