WT-13227 Remove directio adjacent buffer alignment code (#10836)
After investigating it seemed like a good idea to keep the usage of
scratch buffers despite no longer needing it for alignment; scratch
buffers are used in multiple places of our code, and is a convenient way
for us to get a buffer in WT code. Instead we have removed alignment
relating to direct io in the scratch buffer code itself, and made sure
that no other functions using a scratch buffer are relying on this
alignment.  We moved any other checks and alignment relating to scratch
buffers + direct io here too. WT_ITEM_ALIGN was also removed. 