WT-13201 Create a TSan testing variant in evergreen that doesn't fail on error (#10839)
This PR adds a new buildvariant for running TSan tests, but ignoring all
warnings raised by TSan.
This is to catch any compilation or logic errors that might be
introduced by changes that are behind an `#ifdef TSAN_BUILD`.
As tests get added to the `ubuntu2004-tsan` variant we will remove them
from this `no-throw` variant.

---------

Co-authored-by: Luke Pearson <luke.pearson@mongodb.com>