WT-13241 Mark obsolete time window pages dirty during checkpoint cleanup (#10883)
The time window written to the pages to support MVCC can become obsolete
and needs to be cleaned to reclaim the disk space occupied by the
obsolete time window.

As part of the checkpoint cleanup that goes through the obsolete time
window pages in memory, mark them dirty to let the reconciliation remove
the obsolete time window from the pages.