WT-13176 Do not dump txn info if its snapshot has been released (#10762)
do not dump txn info that it's snapshot have been release.


![image](https://github.com/wiredtiger/wiredtiger/assets/14356223/8a7d8267-a73b-4525-a41e-a0dec0c7f4d4)

When I simulated an online problem, I found that the dump snapshot
information was too much, including snapshots that had already been
released, and too much useless information was not conducive to
analyzing the problem.


In addition, I think the following line of code might be better to
remove(I'm not sure if there are any side effects after deletion. I feel
that this line of code placed here is easy to cause txn state dirty
data. for example, txn->flag), because the
（__wt_txn_commit->__txn_release->__wt_txn_release_snapshot) function has
more complete transaction resource release logic.

https://github.com/wiredtiger/wiredtiger/blob/develop/src/txn/txn.c#L1785