WT-12096 Don't validate timestamp against the on-page value if the update is globally visible (#10905)
If an update is globally visible it can be evicted and EBUSY should not
be returned. Thus the check for EBUSY should be skipped if an update is
globally visible. The bug is the check for EBUSY was not skipped if an
update is globally visible.

This was backed out since it made the occurrence of WT-12602 more
frequent. Now it is no longer necessary to wait for WT-12602 and some
BFs need this. So it is time to put this back. The original PR was
#10595.

There is bug fixing and code readability improvements included beyond
the original PR.

---------

Co-authored-by: Chenhao Qu <chenhao.qu@mongodb.com>