WT-13260 Create shared queue mechanism for code coverage (#10914)
### Summary
Create a shared queue mechanism for the parallel work inside code
coverage. The advantage is that all created parallel processes would
share the load of the tests in the code coverage.

### Detailed Description
Previously the coverage report code would bucket all the tests into
their own build directories, and then parallel processes would run them
in the bucket. The tests are separated from the top-down and are no
specific order. The **change** here is that we create a shared queue
mechanism to share the load, with the hopes of equally distributing the
load. This will overall improve the coverage report performance.

Technical details of the code:
- I used a multiprocessing queue that contains all of the build
directories. I used the queue as a shared object to be passed into each
created process and pass it into an initializer function. The function
would then pop off the queue and change their running directory to the
build directory.
- Now the run task function will only run one task at a time instead of
looping through the task list. Also I changed to use the mechanism
executor submit. `executor.submit` pushes the task into it's own queue
mechanism inside `concurrent.futures` and therefore covers the shared
queue mechanism portion of the ticket.
- I created new analysis mode, which grabs the timing of each task and
writes out the tasks in order of timing in descending order into the
same json file. The analysis mode is expected to only be run locally for
now. The reason is behind the new mode is that it would improve the
performance of the shared queue mechanism. Otherwise It is possible that
a
- Per-test code coverage now doesn't have its index build copy as it's
name anymore. I pass in the test number and add it into the process, and
use that instead of the index.
- I separated out the `task_bucket_info` into `task_list` and
`build_dir_list` as it shouldn't be coupled anymore.