WT-12602 Correctly deal with non-timestamped updates in history store (#10686)
The funciton `__hs_delete_reinsert_from_pos `was not properly accounting
for non-timestamped updates and we hit an EBUSY error when we are not
supposed to. In fixing this change, it made sense to refactor the
functionality so that we do not do reinserts unnecessarily, and have
broken out the function so that we can get the cursor position and
counter separately.

---------

Co-authored-by: Jie Chen <jie.chen@mongodb.com>