WT-13282 Add a mechanism to test/model to handle (and skip over) known bugs (#10933)
Add a mechanism for ignoring known bugs and implemented it for WT-13232.
The model can now throw an exception if it knows that it's going to
trigger a known bug, which then results in the rest of that test being
skipped.

This approach works well for the class of bugs that can be modeled
during workload execution, such as WT-13232. I chose this approach as it
provides a simple mechanism to unblock this project. In the future, we
can create the alternative (but much more complicated) mechanism
described in WT-13282, which would allow us to skip over bugs that are
harder to model. We can come back to it in the future if it becomes
necessary.

(cherry picked from commit 385a76ad106eaa278caa36017804b39e7175d370)