WT-13351 Increase the length of test/model runs (#10936)
Increase the length of test/model runs to better cover the parts of the
code that we currently hit with a smaller probability. In particular, we
need this to get a more accurate code coverage runs to see what parts of
RTS we still do not cover.

This change brings up the code coverage report to 90% lines.