WT-12642 Save the connection and table config to the workload file (#10943)
This PR modifies test/model to write WiredTiger's connection and table
configurations to the workload file.

The PR implements this via new `wt_config` workload operation. The
model's own workload runner simply ignores this operation. WiredTiger
workload runner implements the operation by setting the appropriate
fields within its `shared_state` and then uses them when opening a
connection or creating a table. This is stored in the `shared_state`, so
that these configurations could be updated during the workload
execution, although the workload generator does not currently use this
feature.

When the user specifies connection or table config in `tools/model_test`
(the main test program), e.g., via the `-C` and `-T` command-line
arguments, the behavior is as follows:
* If running it with the workload generator: Save the updated
configuration to the workload file.
* If running a workload from a file (using the `-w` argument): Override
the configuration in the workload file.

Note that counterexample reduction is currently allowed to remove the
`wt_config` operations from the workload file. This is intentional so
that the test can further minimize the workload if the given
configuration setting is not necessary.