WT-13352 Add recovery checkpoints that contain evicted prepared data to test/model (#10947)
This PR increases the probability that RTS in the test would see a
checkpoint with evicted prepared data. It accomplishes this by adding a
delay between when a transaction is prepared and when it is committed or
rolled back. It implements this delay by modifying
`kv_workload_generator::generate_transaction` to add a number of no-ops
(`model::operation::nop`) after `prepare_transaction`, which ultimately
causes the commit or rollback operation to come up later in the
execution schedule. (The `nop` operation is not actually added to the
final workload, but I still implemented it as an actual workload
operation as that was more straightforward than doing something
one-off).

After introducing these changes, the PR found a new way to reproduce
WT-13232 for which `kv_table::truncate` didn't account, so the PR fixes
that also. These changes also significantly increased the chances of
reproducing WT-13232 in general, to the point that I added a
`try`/`catch` statement for `model::known_issue_exception` to
`test_workload_generator` to prevent spurious test failures.