WT-13375 Fix ref count for cache->walk_tree (#10954)
Functions `__clear_saved_walk_tree` and `__set_saved_walk_tree`
implement a "smart pointer" pattern.
If `cache->walk_tree` is not null, then it carries a use count.

Whenever it's set to a non-null value, the use count is incremented.
Whenever a non-null value is cleared, the use count is decremented.